# 1.1.0.190 # - Warnings fix and Changelog change D:

+++ IA Julia 1.x.x/News.h
@@ -20,7 +20,7 @@ public:
 	struct sNews
 	{
 		char MSG[3][50];
-		float Time;
+		int Time;
 		int Type;
 	};
